+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`1.0.0...main`][1.0.0...main].
+For a full diff see [`2.0.0...main`][2.0.0...main].
+
+## [`2.0.0`][2.0.0]
+
+For a full diff see [`1.0.0...2.0.0`][1.0.0...2.0.0].
 
 ### Changed
 
@@ -61,13 +65,15 @@ For a full diff see [`4e97e14...0.1.0`][4e97e14...0.1.0].
 [0.3.0]: https://github.com/ergebnis/front-matter/releases/tag/0.3.0
 [0.4.0]: https://github.com/ergebnis/front-matter/releases/tag/0.4.0
 [1.0.0]: https://github.com/ergebnis/front-matter/releases/tag/1.0.0
+[2.0.0]: https://github.com/ergebnis/front-matter/releases/tag/2.0.0
 
 [4e97e14...0.1.0]: https://github.com/ergebnis/front-matter/compare/4e97e14...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/front-matter/compare/0.1.0...0.2.0
 [0.2.0...0.3.0]: https://github.com/ergebnis/front-matter/compare/0.2.0...0.3.0
 [0.3.0...0.4.0]: https://github.com/ergebnis/front-matter/compare/0.3.0...0.4.0
 [0.4.0...1.0.0]: https://github.com/ergebnis/front-matter/compare/0.4.0...1.0.0
-[1.0.0...main]: https://github.com/ergebnis/front-matter/compare/1.0.0...main
+[1.0.0...2.0.0]: https://github.com/ergebnis/front-matter/compare/1.0.0...2.0.0
+[2.0.0...main]: https://github.com/ergebnis/front-matter/compare/2.0.0...main
 
 [#2]: https://github.com/ergebnis/front-matter/pull/2
 [#8]: https://github.com/ergebnis/front-matter/pull/8